feat: :technologist: modulated configuration content in the workspace
