+++ frontend/src/pages/workspace/components/settings/General.tsx
@@ -1,0 +1,49 @@
+import { useState } from "react";
+
+const General = () => {
+  const [workspaceName, setWorkspaceName] = useState("My Workspace");
+  return (
+    <div className="space-y-6">
+      {/* Workspace Name */}
+      <div className="bg-white rounded-lg p-6 shadow-sm">
+        <h3 className="text-lg font-medium text-gray-900 mb-4">
+          Workspace name
+        </h3>
+        <div className="flex space-x-3">
+          <input
+            type="text"
+            value={workspaceName}
+            onChange={(e) => setWorkspaceName(e.target.value)}
+            className="flex-1 px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+          />
+          <button className="px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors">
+            Save
+          </button>
+          <button className="px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors">
+            Cancel
+          </button>
+        </div>
+      </div>
+
+      {/* Workspace Logo */}
+      <div className="bg-white rounded-lg p-6 shadow-sm">
+        <h3 className="text-lg font-medium text-gray-900 mb-4">
+          Workspace logo
+        </h3>
+        <div className="flex items-center space-x-4">
+          <div className="w-16 h-16 bg-gradient-to-br from-purple-500 to-pink-500 rounded-lg flex items-center justify-center">
+            <span className="text-white font-bold">SN</span>
+          </div>
+          <button className="px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors">
+            Upload logo
+          </button>
+        </div>
+        <p className="text-sm text-gray-500 mt-2">
+          Recommended size is 256×256px
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default General;
