+++ frontend/src/pages/workspace/components/settings/Profile.tsx
@@ -1,0 +1,62 @@
+import UserAvatar from "../UserAvatar";
+import { useAuth } from "../../../../hooks/useAuth";
+
+const Profile = () => {
+  const { user } = useAuth();
+
+  return (
+    <div className="space-y-6">
+      {/* Profile Info */}
+      <div className="bg-white rounded-lg p-6 shadow-sm">
+        <h3 className="text-lg font-medium text-gray-900 mb-4">
+          Profile Information
+        </h3>
+        <div className="space-y-4">
+          <div className="flex items-center space-x-4 mb-4">
+            <UserAvatar user={{ name: `${user?.firstName} ${user?.lastName}` }} size="lg" />
+            <button className="px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors">
+              Change Avatar
+            </button>
+          </div>
+          <div className="grid grid-cols-2 gap-4">
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                First Name
+              </label>
+              <input
+                type="text"
+                defaultValue={`${user?.firstName}`}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+              />
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Last Name
+              </label>
+              <input
+                type="text"
+                defaultValue={`${user?.lastName}`}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+              />
+            </div>
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Email
+            </label>
+            <input
+              type="email"
+              defaultValue={`${user?.email}`}
+              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+            />
+          </div>
+          <button className="px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors">
+            Save Changes
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Profile;
